refactor : 내 리포트 목록 화면을 제안 카드 UI로 통합
* 받은 제안 목록 화면의 카드·그리드 레이아웃을 내 리포트 목록으로 이관
* 상태 배지에 상담 중·선택 받지 못함 분기 추가
* 상담 중 카드 하단 문구 추가
* 목록 조회 결과에 중복 행 제거 가드 적용
* 제안 수 기준으로 계산되던 전체 건수 표기 제거
* 받은 제안 목록 라우트를 내 리포트 목록으로 리다이렉트 처리
* 사용처가 없어진 리포트 카드·스켈레톤·에러 폴백 컴포넌트 제거
