+++ apps/web/src/app/customer/proposals/page.tsx
@@ -1,10 +1,6 @@
-import type { Metadata } from "next";
-import { ReceivedProposalsBoundary } from "./_components/ReceivedProposalsBoundary";
+import { redirect } from "next/navigation";
 
-export const metadata: Metadata = {
-  title: "받은 제안",
-};
-
+/** 받은 제안 목록은 내 리포트 목록과 같은 데이터를 보여줘 통합됨(이슈 #262). 기존 진입 경로는 유지. */
 export default function ReceivedProposalsPage() {
-  return <ReceivedProposalsBoundary />;
+  redirect("/customer/reports");
 }
