+++ apps/web/src/app/customer/reports/_components/ReportListBoundary.tsx
@@ -1,26 +1,25 @@
 "use client";
 
-import { QueryErrorResetBoundary } from "@tanstack/react-query";
-import { Suspense, useEffect, useState } from "react";
-import { ErrorBoundary } from "react-error-boundary";
-import { ReportListErrorFallback } from "./ReportListErrorFallback";
+import { AsyncBoundary } from "@/shared/ui/AsyncBoundary";
 import { ReportListSkeleton } from "./ReportListSkeleton";
 import { ReportListView } from "./ReportListView";
 
-export function ReportListBoundary() {
-  const [mounted, setMounted] = useState(false);
-  useEffect(() => setMounted(true), []);
-  if (!mounted) return <ReportListSkeleton />;
+const ERROR_MESSAGES = {
+  FORBIDDEN: {
+    title: "접근 권한이 없어요",
+    desc: "본인 리포트만 확인할 수 있어요.",
+  },
+};
 
+export function ReportListBoundary() {
   return (
-    <QueryErrorResetBoundary>
-      {({ reset }) => (
-        <ErrorBoundary onReset={reset} FallbackComponent={ReportListErrorFallback}>
-          <Suspense fallback={<ReportListSkeleton />}>
-            <ReportListView />
-          </Suspense>
-        </ErrorBoundary>
-      )}
-    </QueryErrorResetBoundary>
+    <AsyncBoundary
+      fallback={<ReportListSkeleton />}
+      errorLayout="flow"
+      errorTitle="리포트를 불러오지 못했어요"
+      errorMessages={ERROR_MESSAGES}
+    >
+      <ReportListView />
+    </AsyncBoundary>
   );
 }
